refactoring works with a lot of props
