+++ components/WorkoutForm.jsx
@@ -1,3 +1,7 @@
+import { useState } from "react";
+import useSWR from "swr";
+import { uid } from "uid";
+
 /**
  * 
  * 
@@ -43,7 +47,109 @@ In the create page's handleCreateWorkout function that's where all the database 
 
  */
 
+export default function WorkoutForm({
+  isSubmitting,
+  setIsSubmitting,
+  onSubmit,
+  workoutPreview,
+  setWorkoutPreview,
+  ...props
+}) {
+  // useSWR Handling
+  const { data: exercises, isLoading, error } = useSWR(`/api/exercises`);
+  if (isLoading || error) {
+    return <main>{error ? "error" : "loading"}</main>;
+  }
 
-export default function WorkoutForm({ ...props }) {
-  return <h2>Workout Form</h2>;
+  //// Submit Handling
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+    if (isSubmitting) return;
+    onSubmit(e);
+  };
+
+  const handleSelect = (id, selectedExercise) => {
+    let newExercises = workoutPreview.exercises.map((exercise) =>
+      exercise._id === id
+        ? { ...exercise, exercise: selectedExercise }
+        : exercise
+    );
+
+    setWorkoutPreview({
+      ...workoutPreview,
+      exercises: newExercises,
+    });
+  };
+
+  const handleDeleteExercise = ({ _id }) => {
+    const newExercises = workoutPreview.exercises.filter(
+      (exercise) => exercise._id !== _id
+    );
+    setWorkoutPreview({ ...workoutPreview, exercises: newExercises });
+  };
+
+  const handleAddExercise = () => {
+    const newExercises = [
+      ...workoutPreview.exercises,
+      { _id: uid(), exercise: "" },
+    ];
+    setWorkoutPreview({ ...workoutPreview, exercises: newExercises });
+  };
+
+  const handleNameChange = (event) => {
+    setWorkoutPreview({ ...workoutPreview, workoutName: event.target.value });
+  };
+
+  // JSX
+  return (
+    <form onSubmit={handleSubmit}>
+      <label for="workoutName">Name</label>
+      <input
+        type="text"
+        name="workoutName"
+        placeholder="enter workout name"
+        value={workoutPreview.workoutName}
+        onChange={(event) => handleNameChange(event)}
+      />
+      {workoutPreview.exercises.map((exerciseInWorkout, index) => {
+        return (
+          <div key={index}>
+            <select
+              name={`exercise-${index}`}
+              value={exerciseInWorkout.exercise}
+              onChange={(event) =>
+                handleSelect(exerciseInWorkout._id, event.target.value)
+              }
+              required
+            >
+              <option value="">Select exercise</option>
+              {exercises.map(({ _id, name: exerciseName }) => {
+                return (
+                  <option value={_id} key={_id}>
+                    {exerciseName}
+                  </option>
+                );
+              })}
+            </select>
+            <button
+              type="button"
+              onClick={() => handleDeleteExercise(exerciseInWorkout)}
+            >
+              delete
+            </button>
+          </div>
+        );
+      })}
+      <button type="button" onClick={handleAddExercise}>
+        add exercise
+      </button>
+      <div>
+        <p>————————</p>
+        <button type="submit" disabled={isSubmitting}>
+          {isSubmitting ? "Creating..." : "Submit"}
+        </button>
+        <button type="reset">reset</button>
+      </div>
+    </form>
+  );
 }
